Message cleanup
git-svn-id: file:///home/jj/hercules.svn/trunk@6977 956126f8-22a0-4046-8f4a-272fa8102e63

+++ ecpsvm.c
@@ -2719,7 +2719,7 @@ static void ecpsvm_showstats2(ECPSVM_STAT *ar,size_t count)
     }
     if(unsupcc)
     {
-        WRMSG(HHC01705,"I",unsupcc);
+        WRMSG(HHC01705,"I",(int)unsupcc);
     }
     return;
 }
